+++ devpi_cleaner/cli.py
@@ -19,6 +19,7 @@ def main(args=None):
     parser.add_argument('package_specification', help='The specification of the package version(s) to remove.')
     parser.add_argument('--batch', help='Assume yes on confirmation questions.', action='store_true')
     parser.add_argument('--dev-only', help='Remove only development versions as specified by PEP 440.', action='store_true')
+    parser.add_argument('--version-filter', metavar='REGEX', help='Remove only versions in which the given regular expression can be found.')
     parser.add_argument('--force', help='Temporarily make indices volatile to enable package removal.', action='store_true')
     parser.add_argument('--password', help='The password with which to authenticate.')
     parser.add_argument('--login', help='The user name to user for authentication. Defaults to the user of the indices to operate on.')
@@ -31,7 +32,9 @@ def main(args=None):
 
     try:
         with DevpiClient(args.server, login_user, password) as client:
-            packages = list_packages(client, args.index_spec, args.package_specification, args.dev_only, None)
+            packages = list_packages(
+                client, args.index_spec, args.package_specification, args.dev_only, args.version_filter
+            )
 
             print('Packages to be deleted: ')
             for package in packages: